Update rftLm.R
Originally the `makeRPV` option existed to reduce computational expenditure if it wasn't going to be used. Because most of this is is fairly efficient now I decided the trade off between consistency and RAM usage leaned in favor of not confusing users with potentially inconsistent results. Also, fixed  some of the masking for cortical voxels.